services/yolo-detection/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pi-yolo-detection"
version = "0.1.0"
description = "Pi video surveillance detection service (Python): Ultralytics YOLOv8 — not installed into the Next.js app."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "Pi internal; Ultralytics AGPL-3.0 (or commercial) — review before distribution." }
dependencies = [
  "ultralytics==8.3.113",
  "fastapi==0.115.8",
  "uvicorn[standard]==0.34.0",
  "pydantic==2.10.6",
  "pillow==11.1.0",
  # Ultralytics currently pins numpy <= 2.1.1
  "numpy==2.1.1",
  "opencv-python-headless==4.11.0.86",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pi_yolo_detection"]

[tool.hatch.build.targets.sdist]
include = ["README.md", "scripts", "pi_yolo_detection", "tests"]
